itp4115-EA-FBI012526-/requirements.txt
alembic==1.8.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.6
cssbeautifier==1.14.7
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
djlint==1.19.7
dnspython==2.2.1
dominate==2.7.0
EditorConfig==0.12.3
email-validator==1.3.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
Flask==2.2.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
fonttools==4.38.0
fqdn==1.5.1
gitdb==4.0.10
GitPython==3.1.29
greenlet==2.0.1
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
importlib-metadata==5.1.0
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsbeautifier==1.14.7
json5==0.9.10
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter-server-mathjax==0.2.6
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.2
jupyterlab==3.5.1
jupyterlab-git==0.40.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.5
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbdime==3.1.1
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
plotly==5.11.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
stack-data==0.6.2
tenacity==8.1.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1


traitlets==5.7.1
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.13
visitor==0.1.3
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
